+++ app/src/main/java/org/oppia/app/player/state/hintsandsolution/HintsAndSolutionExplorationManagerListener.kt
@@ -2,6 +2,7 @@ package org.oppia.app.player.state.hintsandsolution
 
 import org.oppia.app.model.State
 
+/** Listener for fetching current exploration state data. */
 interface HintsAndSolutionExplorationManagerListener {
 
   fun onExplorationStateLoaded(
