+++ app/src/main/java/org/oppia/app/topic/questionplayer/HintsAndSolutionQuestionManagerListener.kt
@@ -2,6 +2,7 @@ package org.oppia.app.topic.questionplayer
 
 import org.oppia.app.model.State
 
+/** Listener for fetching current Question state data. */
 interface HintsAndSolutionQuestionManagerListener {
 
   fun onQuestionStateLoaded(
